experiments/EX-002-QWEN-VL/requirements.txt
transformers>=4.49.0
torch
torchvision
accelerate
opencv-python
pillow
numpy
sentencepiece
requests
